Some AI bug fixes - see Description for more
Moved the Y and Z switch code to the Level blueprint - instead of the
Player

Remember that the debug can still be turned on and off at load time by
the setting on each AI object in the level

Created/cloned a BTAI_Patrol_Common Behaviour Tree to be the common BT
for all the rooms. This has the super set of behviours but only shows
the ones that are enabled individually.

Fixed a few bugs - see GIT diffs

Moved the DebugDraw code so that it runs even if bCanSee bCanHear are
not enabled
Protected more Glogs with "if (DebugDrawEnabled)"